+++ docs/sphinx_source/changelog.rst
@@ -1,6 +1,11 @@
 Changelog
 ==========
 
-Jan. 2022
+Feb. 2023
+
+* Added support for the generation of friend trees. For more details, check :ref:`FriendTree Generation`.
+* Added documentation on ntuple and friend production via KingMaker. For more details, check :ref:`KingMaker`.
+
+Jan. 2023
 
 * Added Quantities <-> Shifts mapping to the output files to allow an easier Postprocessing. For more details, check :ref:`Quantity mapping`.